experimentation/requirements_exp.txt
semantic-kernel==1.18.2
asyncio==3.4.3
mysql-connector-python==9.2.0
SQLAlchemy==2.0.37
aiohttp==3.11.11
aiosqlite==0.20.0
chainlit==2.0.603
python-dotenv==1.0.1
pandas==2.2.3